impt changes
First of all, there were some places in the code where the additional variable of pH was not included in the sigma dissimilarity calculation, so I fixed those.

Also, I figured out that many of the NaNs were coming from the Mahalanobis distance being so large that the probability = 1 (or slightly greater) and as a result the qchisq was infinite.

Finally, I edited the code to output not only the sigma dissimilarity, but also the station of the nearest neighbor and the mahalanobis distance.